Fix linter and add tests for nftables manager ruleset manager
